+++ CHANGELOG.md
@@ -49,6 +49,7 @@ AdminNeo 5.5.0
 - ClickHouse: Fix list of tables (by @Marcus-Nightingale)
 - ClickHouse: Fix offset (by @Marcus-Nightingale)
 - ClickHouse: Set default format, support Common Table Expressions (by @polonskiy)
+- ClickHouse: Fix nullable columns
 
 (Ported relevant changes from Adminer 5.4.0-5.4.1.)